requirements.txt
cachetools==5.3.1
Flask==2.3.2
Flask-Minify==0.42
numpy==1.24.3
openai==0.27.7
datasets==2.12.0 
sentence-transformers==2.2.2 
python-dotenv==1.0.0
protobuf
pinecone-client[grpc]
google-cloud-firestore
google-cloud-secret-manager
google-cloud-storage